cast double -> int

+++ src/object/draw-glow-effect.cpp
@@ -20,9 +20,9 @@ void DrawGlowEffect::draw(int x, Bitmap * work){
     double f = fabs(sin(Util::radians(180) * angle / period));
     int alpha = 50;
 
-    int color_r = (Bitmap::getRed(endColor) - Bitmap::getRed(startColor)) * f + Bitmap::getRed(startColor);
-    int color_g = (Bitmap::getGreen(endColor) - Bitmap::getGreen(startColor)) * f + Bitmap::getGreen(startColor);
-    int color_b = (Bitmap::getBlue(endColor) - Bitmap::getBlue(startColor)) * f + Bitmap::getBlue(startColor);
+    int color_r = (int)((Bitmap::getRed(endColor) - Bitmap::getRed(startColor)) * f + Bitmap::getRed(startColor));
+    int color_g = (int)((Bitmap::getGreen(endColor) - Bitmap::getGreen(startColor)) * f + Bitmap::getGreen(startColor));
+    int color_b = (int)((Bitmap::getBlue(endColor) - Bitmap::getBlue(startColor)) * f + Bitmap::getBlue(startColor));
 
     Bitmap::transBlender(color_r, color_g, color_b, alpha);
